requirements.txt
numpy>=1.24.0
scikit-learn>=1.3.0
flask>=3.0.0
requests>=2.31.0

# Semantic chunking
sentence-transformers>=2.2.0

# PDF support
pdfplumber>=0.10.0
pymupdf>=1.23.0

# Anthropic SDK (answer generation + optional PDF vision)
anthropic>=0.25.0

# .env file support
python-dotenv>=1.0.0

